requirements.txt
pandas>=2.1.0
openpyxl>=3.1.2
tqdm>=4.66.0
python-dotenv>=1.0.0
openai>=1.13.3
numpy>=1.25.0
chardet>=5.2.0
